moving permit loa endpoint to application module
