+++ vehicles/src/modules/common/dto/request/pathParam/permitId.path-param.dto.ts
@@ -1,9 +1,8 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import { IsInt, IsPositive } from 'class-validator';
-import { CompanyIdPathParamDto } from './companyId.path-param.dto';
 
-export class PermitIdPathParamDto extends CompanyIdPathParamDto{
+export class PermitIdPathParamDto {
   @ApiProperty({
     description: `Id of the permit.`,
     example: 1,
