+++ vehicles/src/modules/permit-application-payment/application/dto/request/create-permit-loa.dto.ts
