+++ vehicles/src/modules/permit-application-payment/application/profile/application.profile.ts
@@ -27,6 +27,10 @@ import {
   convertCaseStatus,
 } from '../../../../common/enum/case-status-type.enum';
 import { ReadCaseActivityDto } from '../../../case-management/dto/response/read-case-activity.dto';
+import { CreatePermitLoaDto } from '../dto/request/create-permit-loa.dto';
+import { PermitLoa } from '../entities/permit-loa.entity';
+import { ReadPermitLoaDto } from '../dto/response/read-permit-loa.dto';
+import * as dayjs from 'dayjs';
 
 @Injectable()
 export class ApplicationProfile extends AutomapperProfile {
@@ -441,6 +445,140 @@ export class ApplicationProfile extends AutomapperProfile {
           (d) => d.permitData.permitDataId,
           mapWithArguments((updateApplicationDto, { permitDataId }) => {
             return permitDataId;
+          }),
+        ),
+      );
+      createMap(
+        mapper,
+        CreatePermitLoaDto,
+        PermitLoa,
+        forMember(
+          (d) => d.permitId,
+          mapWithArguments((_, { permitId }) => {
+            return permitId;
+          }),
+        ),
+        forMember(
+          (d) => d.loa.loaId,
+          mapFrom((s) => {
+            return s.loaId[0];
+          }),
+        ),
+        forMember(
+          (d) => d.createdUserGuid,
+          mapWithArguments((_, { userGUID }) => {
+            return userGUID;
+          }),
+        ),
+        forMember(
+          (d) => d.createdUser,
+          mapWithArguments((_, { userName }) => {
+            return userName;
+          }),
+        ),
+        forMember(
+          (d) => d.createdUserDirectory,
+          mapWithArguments((_, { directory }) => {
+            return directory;
+          }),
+        ),
+
+        forMember(
+          (d) => d.createdDateTime,
+          mapWithArguments((_, { timestamp }) => {
+            return timestamp;
+          }),
+        ),
+
+        forMember(
+          (d) => d.updatedUserGuid,
+          mapWithArguments((_, { userGUID }) => {
+            return userGUID;
+          }),
+        ),
+        forMember(
+          (d) => d.updatedUser,
+          mapWithArguments((_, { userName }) => {
+            return userName;
+          }),
+        ),
+        forMember(
+          (d) => d.updatedUserDirectory,
+          mapWithArguments((_, { directory }) => {
+            return directory;
+          }),
+        ),
+
+        forMember(
+          (d) => d.updatedDateTime,
+          mapWithArguments((_, { timestamp }) => {
+            return timestamp;
+          }),
+        ),
+      );
+
+      createMap(
+        mapper,
+        PermitLoa,
+        ReadPermitLoaDto,
+        forMember(
+          (d) => d.permitId,
+          mapFrom((s) => {
+            return s.permitId;
+          }),
+        ),
+        forMember(
+          (d) => d.permitLoaId,
+          mapFrom((s) => {
+            return s.permitLoaId;
+          }),
+        ),
+        forMember(
+          (d) => d.loa.loaId,
+          mapFrom((s) => {
+            return s.loa.loaId;
+          }),
+        ),
+        forMember(
+          (d) => d.loa.companyId,
+          mapFrom((s) => {
+            return s.loa.company.companyId;
+          }),
+        ),
+        forMember(
+          (d) => d.loa.startDate,
+          mapFrom((s) => {
+            return dayjs(s.loa.startDate).format('YYYY-MM-DD');
+          }),
+        ),
+        forMember(
+          (d) => d.loa.expiryDate,
+          mapFrom((s) => {
+            if (s.loa.expiryDate) return dayjs(s.loa.expiryDate).format('YYYY-MM-DD');
+          }),
+        ),
+        forMember(
+          (d) => d.loa.loaPermitType,
+          mapFrom((s) => {
+            return s.loa.loaPermitTypes.map((lpt) => lpt.permitType);
+          }),
+        ),
+        forMember(
+          (d) => d.loa.powerUnits,
+          mapFrom((s) => {
+            if (s.loa.loaVehicles)
+              return s.loa.loaVehicles
+                .filter((lv) => lv.powerUnit)
+                .map((lv) => lv.powerUnit);
+          }),
+        ),
+        forMember(
+          (d) => d.loa.trailers,
+          mapFrom((s) => {
+            if (s.loa.loaVehicles)
+              return s.loa.loaVehicles
+                .filter((lv) => lv.trailer)
+                .map((lv) => lv.trailer);
           }),
         ),
       );
